(feat) Remove server config file and provide example
