+++ distribution/server/config/config.example.js
@@ -1,0 +1,8 @@
+var server = require('../server.js');
+var config = {};
+
+config.mongoURL = 'mongodb://localhost/dropnet';
+
+config.localPort = 3000;
+
+module.exports = config;
